pyproject.toml
[project]
name = "openflow"
version = "0.2.0"
description = "Open-source n8n-style workflow engine with visual LangGraph agent nodes"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.9.0",
  "langgraph>=0.2.0",
  "langchain-ollama>=0.2.0",
  "tavily-python>=0.5.0",
  "pyyaml>=6.0.0"
]

[tool.uv]
dev-dependencies = []
